Add files via upload
This document contains trimestral data history of three companies. 
NVIDIA, AMD and INTEL.  This information was recolected in their trimestral financial CFO reports visible in their own website and the SEC (Securty exchange comission).

This information is about their financial reports I considered important to do the analysis such as:
Stock price (last). Revenue. Gross Margin. Net Income.	Data Center. Free cashflow. The Data center column are the main sales of NVIDIA, this is so IMPORTANT because is related to the infrastructure sold to third parties which are looking to develop and mantain AI. AI is changing the world in every field aswell as changing every person lifestyle.