Add files via upload
Aiming to discover the most important factors that affects NVIDIA stock price I made use of python to excecute a random forest and linear regression. The results of R2 are 0,8 for random forest which means the model fits very well the real data. Here you will have the results but not the code, however, If you want the code to excecute the random forest and linear regression please do not hesitate to contact me.