Add files via upload
Add files via upload
This power bi document contains the resume of the NVDIA Stock price analysis from 2022 to 2024, going from its financials, returns vs risks, factors that most affect the stock price, lastly, a prediction of its price in the future using an Arima model. To start with, you will be able to compare the financials of NVDIA, AMD and INTEL going from net income, free cash flow, stock price evoluction and more. Secondly, you will have in hand the sharpe ratio for these three companies and see which one gives better returns compared to its volatility, the highest value is the best. Thirdly, a random forest regression was performed to see the main variables that affects NVDIA stock price. Lastly and not least important, an Arima model was excuted to try to predict the future of the NVIDIA stock price. As a result, on one hand I get the result that its important for an investor to take a look on the evolution of AMD stock price and the sales of NVDIA of Data center if these ones are increasing the price might get to $170 per share at the end of the year. On the other hand this research has a limit, its limit is the sentiment analysis which can help an investor to take care every day of his or her investment on NVIDIA. The NVDIA stock price has been increasing since the start of the year 2025 to $140 as this model predicted. Nonetheless, the price decreased on february as NVIDIA was removed from 3 BIllionare group and this model couldn't anticipate this actions, but, along with a sentiment analysis, an investor might have a good opportunity to increase his or her wealth and reduce risks. Thank you for your time, if you have any opinion or questions I'm here for you.